backend/requirements.txt
alembic==1.13.2
beautifulsoup4==4.13.4
chromadb==1.0.12
email-validator==2.2.0
fastapi==0.115.9
numpy==1.26.4
openai==1.86.0
openpyxl==3.1.5
orjson==3.10.18
pandas==2.3.0
passlib[bcrypt]==1.7.4
pillow==11.2.1
playwright==1.52.0
posthog==4.7.0
psycopg[binary]==3.2.1
pydantic==2.11.5
pytest==8.4.1
python-dotenv==1.1.0
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
scikit-learn==1.7.0
sentence-transformers==4.1.0
SQLAlchemy==2.0.36
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
ultralytics==8.3.153
uvicorn==0.34.3
allure-pytest
pytest-html
duckdb
matplotlib
plotly==5.24.1
axe-playwright-python
behave==1.2.6
allure-behave